more int32_t

+++ src/ariadne.h
@@ -171,7 +171,7 @@ struct Ariadne { // sizeof = 102
   struct Coord3d manoeuvre_requested_position;
   unsigned char manoeuvre_state;
   short wallhug_angle;
-  long straight_dist_to_next_waypoint;
+    int32_t straight_dist_to_next_waypoint;
 };
 
 struct PathWayPoint { // sizeof = 8
@@ -219,8 +219,8 @@ struct Navigation {
   unsigned char wallhug_retry_counter;
   unsigned char wallhug_state;
   unsigned char push_counter;
-  long dist_to_final_pos;
-  long distance_to_next_pos;
+    int32_t dist_to_final_pos;
+    int32_t distance_to_next_pos;
   int32_t angle;
   SubtlCodedCoords first_colliding_block;
   SubtlCodedCoords second_colliding_block;
