+++ src/power_hand.h
@@ -116,7 +116,7 @@ struct HandRule {
     char type;
     char enabled;
     char allow; // allow: 1, deny: 0
-    long param;
+    int32_t param;
 };
 
 TbBool eval_hand_rule_for_thing(struct HandRule *rule, const struct Thing *thing_to_pick);
